+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 1.0 (unreleased)
 ----------------
 
-- Use (-100, 1) years range for birthday field.
+- Use (-200, 1) years range for birthday field.
   [vincentfretin]
 
 - "Add new" popup link is renamed from "Add ..." to "Create ..."